LayoutTests/fast/forms/validation-message-maxLength-expected.txt

Tests the validation message when maxlength is used.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "abc"
PASS input.checkValidity() is false
PASS input.validationMessage is "Use no more than 2 characters"
PASS input.value is "ab"
PASS input.checkValidity() is false
PASS input.validationMessage is "Use no more than one character"
PASS input.value is "a"
PASS input.checkValidity() is false
PASS input.validationMessage is "Use no more than 0 characters"
PASS successfullyParsed is true

TEST COMPLETE
